requirements.txt
gitpython>=3.1.0
pyyaml>=5.1
requests>=2.25.0
semgrep>=1.21.0
openai>=1.3.0 
